Implemented custom @Name validation

+++ src/main/java/com/telerikacademy/virtualteacher/repositories/UserRepository.java
@@ -11,5 +11,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     @Modifying
     @Query(value = "update users u set u.enabled = 0 where u.user_id = :id", nativeQuery = true)
-    void deleteById(@Param("id") Integer id);
+    void deleteById(@Param("id") Long id);
 }
